+++ src/hooks/useTimeStamp.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 
 export default function useTimeStamp(timestamp: string): string {
-  const [timeState, setTimeState] = useState("");
+  const [timeState, setTimeState] = useState<string>("");
 
   const updateTimeStamp = () => {
     const timeAgo = Math.floor(
@@ -26,25 +26,6 @@ export default function useTimeStamp(timestamp: string): string {
       const years = Math.floor(timeAgo / (60 * 60 * 24 * 7 * 30));
       setTimeState(`${years}년 전`);
     }
-  };
-
-  useEffect(() => {
-    updateTimeStamp();
-  }, [timestamp]);
-
-  return timeState;
-}
-
-export function useCalDay(timestamp: string): string {
-  const [timeState, setTimeState] = useState("");
-
-  const updateTimeStamp = () => {
-    const timeLeft = Math.floor(
-      (new Date(timestamp).getTime() - new Date().getTime()) / 1000
-    );
-
-    const days = Math.floor(timeLeft / (60 * 60 * 24));
-    setTimeState(`D-${days}`);
   };
 
   useEffect(() => {
